requirements.txt
python-telegram-bot[ext]
feedparser
stravalib
Flask
requests
apscheduler
httpx
pytz
python-dotenv